requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Azure dependencies
azure-storage-blob>=12.19.0  # Azure Blob Storage support
azure-identity>=1.15.0  # Azure authentication

# Database drivers
# PostgreSQL support (required for database storage backend)
psycopg2-binary>=2.9.0

# Legacy database support (optional, not used by current pipeline)
# MySQL/MariaDB support
# mysql-connector-python>=8.0.0
